% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tocestatfr20250927}
\alias{tocestatfr20250927}
\title{TOC Eurostat 2025-09-27}
\format{
A data frame with x rows and y variables:
\describe{
\item{title}{title}
\item{code}{code}
\item{type}{type}
\item{last.update.of.data}{last update of data}
\item{last.table.structure.change}{last table structure change}
\item{data.start}{data start}
\item{data.end}{data end}
\item{values}{values}
\item{hierarchy}{hierarchy}
}
}
\usage{
tocestatfr20250927
}
\description{
Dataset contenant des infos sur les donnees dispo sur eurosat
Les variables sont les suivantes
}
\keyword{datasets}
